extracted out Product class & interfaces
